backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic[email]==2.5.0
python-multipart==0.0.6
langchain==0.1.0
langchain-openai==0.0.5
openai==1.3.0
python-dotenv==1.0.0
motor==3.3.2
pymongo==4.6.0
httpx==0.25.2
